svg/painting/parsing/fill-rule-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: getComputedStyle().fillRule</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#FillRuleProperty" rel="help"/><h:meta content="fill-rule computed value is as specified." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("fill-rule", "nonzero");
test_computed_value("fill-rule", "evenodd");

  ]]></script></svg>
